update coin market. use bigint for max_supply

+++ sql/create_tbs.sql
@@ -32,8 +32,8 @@ CREATE TABLE if not exists `coin_list` (
 CREATE TABLE if not exists `coin_markets` (
   `id` char(100) COLLATE utf8_bin NOT NULL COMMENT 'id',
   `symbol` char(40) COLLATE utf8_bin NOT NULL COMMENT 'symbol',
-  `total_supply` int(20) NOT NULL COMMENT '总市值',
-  `max_supply` int(20) NOT NULL COMMENT '流通市值',
+  `total_supply` bigint NOT NULL COMMENT '总市值',
+  `max_supply` bigint NOT NULL COMMENT '流通市值',
   `max_supply_updated_hour_ts` int(11) NOT NULL COMMENT '流通市值更新小时时间戳',
   PRIMARY KEY (`id`, `symbol`, `max_supply`)
 );
